Quarter 05/Q5_Hackathon-01_Customer_Support_Agent/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
langchain-google-genai
psycopg2-binary